tests/wpt/web-platform-tests/scroll-animations/css/at-scroll-timeline-orientation.html

<!DOCTYPE html>
<link rel="help" src="https://drafts.csswg.org/scroll-animations-1/#scroll-timeline-at-rule">
<link rel="help" src="https://drafts.csswg.org/scroll-animations-1/#descdef-scroll-timeline-orientation">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/web-animations/testcommon.js"></script>
<style>
  #scroller {
    overflow: scroll;
    width: 100px;
    height: 100px;
  }
  #contents {
    height: 300px;
    width: 300px;
  }
  @keyframes expand {
    from { width: 100px; }
    to { width: 200px; }
  }
  @scroll-timeline timeline_auto {
    source: selector(#scroller);
    orientation: auto;
    time-range: 10s;
    start: 0px;
    end: 100px;
  }
  @scroll-timeline timeline_vertical {
    source: selector(#scroller);
    orientation: vertical;
    time-range: 10s;
    start: 0px;
    end: 100px;
  }
  @scroll-timeline timeline_horizontal {
    source: selector(#scroller);
    orientation: horizontal;
    time-range: 10s;
    start: 0px;
    end: 100px;
  }
  @scroll-timeline timeline_block {
    source: selector(#scroller);
    orientation: block;
    time-range: 10s;
    start: 0px;
    end: 100px;
  }
  @scroll-timeline timeline_inline {
    source: selector(#scroller);
    orientation: inline;
    time-range: 10s;
    start: 0px;
    end: 100px;
  }
  #container > div {
    width: 0px;
    animation: expand 10s linear;
  }
  /* Ensure stable expectations if feature is not supported */
  @supports not (animation-timeline:foo) {
    #container > div { animation-play-state: paused; }
  }
  .horizontal { writing-mode: horizontal-tb; }
  .vertical { writing-mode: vertical-lr; }
  #element_auto { animation-timeline: timeline_auto; }
  #element_vertical { animation-timeline: timeline_vertical; }
  #element_horizontal { animation-timeline: timeline_horizontal; }
  #element_block_in_horizontal { animation-timeline: timeline_block; }
  #element_inline_in_horizontal { animation-timeline: timeline_inline; }
  #element_block_in_vertical { animation-timeline: timeline_block; }
  #element_inline_in_vertical { animation-timeline: timeline_inline; }
</style>
<div id=scroller>
  <div id=contents></div>
</div>
<div id=container>
  <div id=element_auto></div>
  <div id=element_vertical></div>
  <div id=element_horizontal></div>
  <div id=element_block_in_horizontal class="horizontal"></div>
  <div id=element_inline_in_horizontal class="horizontal"></div>
  <div id=element_block_in_vertical class="vertical"></div>
  <div id=element_inline_in_vertical class="vertical"></div>
</div>
<script>
  // Animations linked to a vertical scroll-timelines are at 75% progress.
  scroller.scrollTop = 75;
  // Animations linked to a horizontal scroll-timelines are at 25% progress.
  scroller.scrollLeft = 25;

  promise_test(async (t) => {
    await waitForNextFrame();
    assert_equals(getComputedStyle(element_auto).width, '175px');
  }, 'Orientation auto behaves as expected');

  promise_test(async (t) => {
    await waitForNextFrame();
    assert_equals(getComputedStyle(element_vertical).width, '175px');
  }, 'Orientation vertical behaves as expected');

  promise_test(async (t) => {
    await waitForNextFrame();
    assert_equals(getComputedStyle(element_horizontal).width, '125px');
  }, 'Orientation horizontal behaves as expected');

  promise_test(async (t) => {
    await waitForNextFrame();
    assert_equals(getComputedStyle(element_block_in_horizontal).width, '175px');
  }, 'Orientation block behaves as expected in horizontal writing-mode');

  promise_test(async (t) => {
    await waitForNextFrame();
    assert_equals(getComputedStyle(element_inline_in_horizontal).width, '125px');
  }, 'Orientation inline behaves as expected in horizontal writing-mode');

  promise_test(async (t) => {
    await waitForNextFrame();
    assert_equals(getComputedStyle(element_block_in_vertical).writingMode, 'vertical-lr');
    assert_equals(getComputedStyle(element_block_in_vertical).width, '125px');
  }, 'Orientation block behaves as expected in vertical writing-mode');

  promise_test(async (t) => {
    await waitForNextFrame();
    assert_equals(getComputedStyle(element_inline_in_vertical).writingMode, 'vertical-lr');
    assert_equals(getComputedStyle(element_inline_in_vertical).width, '175px');
  }, 'Orientation inline behaves as expected in vertical writing-mode');

</script>
