chromium/third_party/blink/web_tests/external/wpt/scroll-animations/css/at-scroll-timeline-orientation-expected.txt

This is a testharness.js-based test.
FAIL Orientation auto behaves as expected assert_equals: expected "175px" but got "100px"
FAIL Orientation vertical behaves as expected assert_equals: expected "175px" but got "100px"
FAIL Orientation horizontal behaves as expected assert_equals: expected "125px" but got "100px"
FAIL Orientation block behaves as expected in horizontal writing-mode assert_equals: expected "175px" but got "100px"
FAIL Orientation inline behaves as expected in horizontal writing-mode assert_equals: expected "125px" but got "100px"
FAIL Orientation block behaves as expected in vertical writing-mode assert_equals: expected "125px" but got "100px"
FAIL Orientation inline behaves as expected in vertical writing-mode assert_equals: expected "175px" but got "100px"
Harness: the test ran to completion.
